Refactoring 6/18/24 @ 14:57

+++ Authentication/AuthenticationLogout.php
@@ -2,8 +2,12 @@
 
 namespace SEVEN_TECH\Gateway\Authentication;
 
+use SEVEN_TECH\Gateway\Cookie\Cookie;
 use SEVEN_TECH\Gateway\Database\DatabaseExists;
 use SEVEN_TECH\Gateway\Exception\DestructuredException;
+use SEVEN_TECH\Gateway\Services\Redis\RedisSession;
+use SEVEN_TECH\Gateway\Session\Session;
+use SEVEN_TECH\Gateway\Session\SessionRedis;
 use SEVEN_TECH\Gateway\Session\SessionWordpress;
 use SEVEN_TECH\Gateway\Token\Token;
 use SEVEN_TECH\Gateway\Token\TokenFirebase;
@@ -27,11 +31,11 @@ class AuthenticationLogout
         try {
             (new Validator)->isValidEmail($request['email']);
 
-            $refreshToken = (new Token)->getRefreshToken($request);
-            $verifier = (new Token)->hashToken($refreshToken);
+            $token = (new Cookie($_COOKIE))->token;
+            $verifier = (new Token)->hashToken($token);
 
-            $session_destroyed = (new SessionWordpress)->deleteSession($request['id'], $verifier);
-// delete redis session
+            $session_destroyed = (new Session)->deleteSession($request['id'], $verifier);
+
             if (!$session_destroyed) {
                 throw new Exception('Unable to remove session.');
             }
@@ -60,20 +64,33 @@ class AuthenticationLogout
     public function logoutAll(WP_REST_Request $request)
     {
         try {
-            (new DatabaseExists)->existsByEmail($request['email']);            
-            
+            (new DatabaseExists)->existsByEmail($request['email']);
+
+            $this->tokenFirebase->revokeAllRefreshTokens($request);
+
             if (!isset($request['id'])) {
                 throw new Exception('ID is required to logout all accounts.', 500);
             }
 
-            $session_tokens_deleted = delete_user_meta($request['id'], 'session_tokens');
+            $wordpresSessions = (new SessionWordpress)->getSessions($request['id']);
 
-            if (!$session_tokens_deleted) {
-                throw new Exception('There was an error deleting sessions.', 500);
+            if (is_array($wordpresSessions)) {
+                $session_tokens_deleted = delete_user_meta($request['id'], 'session_tokens');
+
+                if (!$session_tokens_deleted) {
+                    throw new Exception('There was an error deleting sessions.', 500);
+                }
             }
 
-            $this->tokenFirebase->revokeAllRefreshTokens($request);
-// delete redis session
+            if ((new RedisSession)->isReady) {
+                $redisSessions = (new SessionRedis)->getSessions($request['id']);
+
+                if (is_array($redisSessions)) {
+                    foreach ($redisSessions as $key => $value) {
+                        (new SessionRedis)->deleteSession($key);
+                    }
+                }
+            }
 
             wp_logout();
